requirements.txt
groq
python-dotenv
sqlalchemy
psycopg2-binary
rich
typer[all]
markdown
weasyprint

